Integrate trails color scheme into base theme
