+++ .changeset/update-theme.md
@@ -1,0 +1,6 @@
+---
+"@open-pioneer/base-theme": minor
+---
+
+Merge developments from @open-pioneer/theme.
+This package now provides the `"trails"` color scheme, among other things.
