--- src/packages/theme/CHANGELOG.md
@@ -1,19 +1,0 @@
-# @open-pioneer/theme
-
-## 0.3.0
-
-### Minor Changes
-
--   ee7c2d4: Update runtime version.
-
-## 0.2.0
-
-### Minor Changes
-
--   70349a8: Update to new core packages major versions
-
-## 0.1.0
-
-### Minor Changes
-
--   2cee00c: Initial release.
