--- src/packages/theme/build.config.mjs
@@ -1,11 +1,0 @@
-// SPDX-FileCopyrightText: 2023 Open Pioneer project (https://github.com/open-pioneer)
-// SPDX-License-Identifier: Apache-2.0
-
-import { defineBuildConfig } from "@open-pioneer/build-support";
-
-export default defineBuildConfig({
-    entryPoints: ["index"],
-    publishConfig: {
-        strict: true
-    }
-});
