--- src/packages/theme/index.ts
@@ -1,3 +1,0 @@
-// SPDX-FileCopyrightText: 2023 Open Pioneer project (https://github.com/open-pioneer)
-// SPDX-License-Identifier: Apache-2.0
-export { theme } from "./theme";
