--- src/packages/theme/package.json
@@ -1,21 +1,0 @@
-{
-    "name": "@open-pioneer/theme",
-    "version": "0.3.0",
-    "license": "Apache-2.0",
-    "main": "index.ts",
-    "scripts": {
-        "build": "build-pioneer-package"
-    },
-    "peerDependencies": {
-        "@open-pioneer/chakra-integration": "workspace:^",
-        "@open-pioneer/runtime": "workspace:^",
-        "@open-pioneer/base-theme": "workspace:^"
-    },
-    "devDependencies": {
-        "core-packages": "workspace:^"
-    },
-    "publishConfig": {
-        "directory": "dist",
-        "linkDirectory": false
-    }
-}
