--- src/packages/theme/typedoc.json
@@ -1,4 +1,0 @@
-{
-    "extends": ["core-packages/typedoc.base.json"],
-    "entryPoints": ["./index.ts"]
-}
