+++ src/samples/auth-sample/auth-app/package.json
@@ -7,7 +7,6 @@
         "@open-pioneer/chakra-integration": "workspace:^",
         "@open-pioneer/core": "workspace:^",
         "@open-pioneer/runtime": "workspace:^",
-        "@open-pioneer/theme": "workspace:^",
         "react": "^18.2.0"
     }
 }
