+++ src/samples/notify-sample/notify-app/app.ts
@@ -1,13 +1,11 @@
 // SPDX-FileCopyrightText: 2023 Open Pioneer project (https://github.com/open-pioneer)
 // SPDX-License-Identifier: Apache-2.0
 import { createCustomElement } from "@open-pioneer/runtime";
-import { theme } from "@open-pioneer/theme";
 import * as appMetadata from "open-pioneer:app";
 import { AppUI } from "./AppUI";
 
 const Element = createCustomElement({
     component: AppUI,
-    theme,
     appMetadata
 });
